style(ruff): widen rule set and lock import discipline
Builds on the baseline configured in phase 3.1 by enabling four extra
rule families and tightening import hygiene. The whole project (and the
new tests/) stays lint-clean.

Newly enabled rule sets:
* PIE  — flake8-pie (misc structural lints).
* RET  — flake8-return (return idioms).
* RUF  — ruff-specific rules.
* PT   — flake8-pytest-style.
* TID  — flake8-tidy-imports (relative imports are now banned).

isort is told that `macmanager` and `tests` are first-party, so the
import groups stay deterministic across editors.

Targeted ignores that survive the rule expansion:
* RUF001/002/003 — ambiguous unicode in strings/docstrings/comments:
  the project intentionally keeps verbatim non-English error messages
  (e.g. Java's localized "couldn’t be completed") for matching.
* RET504, SIM117 — readability-over-cleverness preferences.
* `tests/**` — relax B011 and PT011 to keep smoke tests pragmatic.
* `scripts/**` — `print` is allowed in throwaway maintenance scripts.

Code fixes the new rules surfaced:
* `ui.py`: drop redundant `int(round(...))` — `round()` already
  returns `int` in Python 3 (RUF046).
* `watch.py`: collapse two consecutive `from macmanager.security`
  imports into a single sorted block (I001).

`ruff check . && ruff format --check . && pytest` all green.

Closes phase 5.1 of the open-source hardening roadmap.

+++ macmanager/ui.py
@@ -51,6 +51,6 @@ def usage_color(pct: float, warn: float = 70, crit: float = 90) -> str:
 def bar(pct: float, width: int = 20, color: str | None = None) -> str:
     """Simple text-based progress bar using rich markup."""
     pct = max(0.0, min(100.0, pct))
-    filled = int(round(width * pct / 100))
+    filled = round(width * pct / 100)
     color = color or usage_color(pct)
     return f"[{color}]{'█' * filled}{'░' * (width - filled)}[/]"
